requirements.txt
# Pydantic AI Core
pydantic-ai>=0.1.0
pydantic>=2.5.0
pydantic-settings>=2.0.0

# Environment Management
python-dotenv>=1.0.0

# LLM Provider
openai>=1.10.0

# Web Scraping Stack
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
fake-useragent>=1.4.0

# Async Support
httpx>=0.26.0
aiofiles>=23.0.0

# CLI Interface
typer>=0.9.0
rich>=13.0.0
click>=8.1.0

# MCP Server Implementation
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic-core>=2.14.0

# Data Processing
pandas>=2.1.0
jsonlines>=4.0.0

# Testing Framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-httpx>=0.25.0
httpx-mock>=0.10.0

# Code Quality
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0

# Enhanced Logging
loguru>=0.7.0
structlog>=23.2.0